feat: Implement showing material list
